Remove superfluous settings and comments
Remove superfluous settings and comments from texcavator settings
(settings.py and settings_local.py). Settings (and comments) removed
refer to:
- settings for specific machines
- checks and settings for older python and django versions

Setting DJANGO_MAJ_MIN is used in urls.py; this must be fixed next.
Also, I left settings for PROJECT_CELERY and PROJECT_BROKER, because at
the moment it is not clear how they are used in the app.