Remove version specific instantiation of url patterns
The url patterns were instantiated based on the version of Django that
is used to run the app. As we agreed to remove support for older
versions of Django (at the moment we use Django 1.5.5), this version
specific instantiation was removed.

Because the version check used a global variable
settings.DJANGO_MAJ_MIN that was previously removed, this also solved
the error caused by removing this global variable.