Link url services/search to new view function search()
Currently, all (or almost all) services are mapped to a big 'proxy'
function that analyzes the urls and calls the appropriate functions.
This is better handled by having a dedictad function in
services/views.py for each of the services offered. This is the first
step: services/search is now mapped to the search() function in
services/views.py. At this moment the search function returns a simple
string in a HttpResponse object. The next steps are to search in the
elasticsearch index and to format the results.