Add new elasticsearch functionality
Added new functionality to perform searches directly in the
elasticsearch index. This functionality uses the elasticsearch python
client. At the moment At the moment, the literal query string typed in
by the user is inserted in the elasticsearch query. Functionality to
handle more complex queries needs to be added, as well as updating the
date filter with the appropriate date range.