Generate single document word cloud with Elasticsearch
The functionality to generate a word cloud for a single document no
longer uses xtas, but retrieves the term vectors from Elasticsearch and
formats the data in the format expected by the user interface:
{
        'status': 'ok'
        'max_count': ...
        'result':
                [
                        {
                                'term': ..
                                'count': ...
                        },
                        ...
                ]
}

Currently, the title of the article is ignored when generating the word
cloud. The title should be included. This requires a field in the
Elasticsearch index that contains both the title and text of articles
and stores the term vectors for this field.

In addition, stop words are not filtered. This should also be done at
index time.