Handle invalid document ids
Added functionality to deal with invalid document ids in the
single_document_word_cloud function. Invalid document ids return an
error that is displayed in the user interface.

Also added test to check the output of single_document_word_cloud with
an invalid document id as input. Still need to add tests to check
output with valid document ids.