Change function test_cql2es to accept any query
The function test_cql2es in services/views.py checks whether a query
string can be translated to an Elasticsearch query by a custom perl
script. The script does not work (it requires a specific version of
perl installed with perlbrew and it does not produce valid
Elasticsearch queries). Therefore, it will be deleted. Pending
deletion, the function has been changed to accept any query; for
empty queries an error message is returned.