Remove cql2es from function doc_count
The doc_count function also used cql2es to translate queries. Because
cql2es will be removed, doc_count was rewritten not to call cql2es. In
addition, the code was cleaned up.

Doc_count (and other functions) make extensive use of returning custom
error messages in json. This is a lot of repetative code that should be
replaced by a custom function.