Generate multiple documents word clouds with Elasticsearch
Word clouds for saved queries (based on a lexicon id) are generated
using an Elasticsearch aggregation query.

The word clouds are based on only the text of documents. Later, the
title should be analyzed as well. At the moment, the 100 words that
occur most frequently are returned. This is a parameter that can be
changed. We plan to run tests to determine whether this parameter
should be changed.

The old code also specifies a word cloud can be generated based on
(NER) tags. This functionality still needs to be implemented. Also, it
might be possible that a word cloud needs to be generated for multiple
document ids. Currently, in this case an error message is returned. It
is not clear whether this possibility is actually used at all.