Fix single document word clouds
Single document word clouds could not be generated, because the code
depended on a function in the xtas module (that was previously
removed). The call to the xtas function was replaced with an
dojo.xhrGet call to xtas/services/cloud. To remove all traces of (old)
xtas, the url to get word clouds should be changed (in
services/urls.py).