Remove validation of xtas key from createQueryLine()
In javascript function createQueryLine() buttons for different
actions on saved queries (including generating the word cloud)
were disabled (and therefore not displayed) because their visibility
depended on the validation of the xtas key. Because xtas is to be
removed from the interface entirely, the xtas key is not validated
anymore and all buttons are always displayed.

The next step is to generate a word cloud without referring to xtas.
Clicking the 'generate word cloud' button now results in an error.