Update functionality to retrieve document ids for a query
Time lines are generated based on custom functionality for bursts
detection (lexicon/burststdetection/bursts.py). This works based on a
list of document ids together with dates. The code for generating
time lines did not work, because it used the cql2es perl script to
translate queries to Elasticsearch. In order to make the code work, the
cql2es translation step was removed and new functionality to get a list
of document ids together with dates was added to services/es.py.

Generating this list seems to work. However, for the generation of time
lines another data structure is required that involves DayStatistics.
It seems that these must be generated before time lines can be
generated.