Add functionality to gather DayStatistics from Elasticsearch
The old gatherstatistics command contained code to retrieve
DayStatistics directly from the KB (Koninklijke Bibliotheek). This code
was not working, so it was replaced with code that gathers
DayStatistics from Elasticsearch. A DayStatistic is the total number of
documents in the Elasticsearch index for a given day.

After running the gatherstatistics command, the time line functionality
works.

We now retrieve the 'raw' number of documents; a normalized value could
also be used. We'll have to experiment to find out what works best.