Remove xtas from time line clouds
When clicking on a bar in a time line, a word cloud should be
displayed. This functionality was not working, because the
javascript function for generating this word cloud called xtas.
The call to xtas was replaced with a direct call to services/cloud.
However, the word cloud is not displayed yet, because the functionality
for word clouds based on multiple document ids is not yet implemented
(an error message stating this is returned). The next step is to
implement word clouds based on multiple document ids.