Change create_query() to accept boolean queries
Elasticsearch supports boolean queries through query string queries.
The code for generating an Elasticsearch query was updated to create a
query string query instead of a boolean query. Parsing the query is
left to Elasticsearch. Because the user might enter an invalid boolean
query, the query should be validated before it is send to Elasticsearch
to get the results.