Validate query before searching
Queries are validated before they are send to Elasticsearch to retrieve
results. If the query is not valid an error message explaining why the
query is invalid is returned. Currently, the error message is displayed
in the interface where the search results are displayed. This should be
changed to showing a proper error message.