Add model to store document ids
To test the ElasticSearch performance on term aggregation queries, we
want to control the number of documents requested in a query. To this
end, document id queries will be used. In order to make document ids
queries, we need a collection of document ids to choose from.
Therefore, a collection of document ids will be stored in the database.
In this commit, the model for storing document ids was added. The next
step is to add a command to gather a collection of document ids from
the ElasticSearch index.