Add admin command to perform ElasticSearch terms aggregations
The admin command esperformance was added. This command performs
ElasticSearch terms aggregations. The number of documents per query and
the number of repetitions (each time with a different set of random
documents) can be specified.
Options that should be added later are: whether the ES cache should be
flushed before executing the terms aggregation, the number of terms to
be retrieved, and the minimal number of documents in which a term
should occur before it is added to the aggregation.