Change researchSubmit() to use searchSubmit()
The javascript function researchSubmit() is called when the research
button of a saved query is clicked. researchSubmit() was a copy of the
old searchSubmit() that was changed in the previous commit. Because of
this, the results were no longer properly displayed when the re-search
button was clicked. Therefore, researchSubmit() was changed to call
searchSubmit().