Remove services/moses from back-end
Moses is a statistical machine translation system
(http://www.statmt.org/moses/) that at some time was used in Biland(?).
Texcavator deals with a single language, so no translation is required.
Therefore, Moses was removed from the back-end. The next step is to
also remove it from the front-end. Also, in settings_local.py there is
a setting for Google Translate that needs to be removed.