Remove DUAL_MODE from front-end
To determine whether translation options are made available in the
front-end, the variable DUAL_MODE was set. Since Texcavator does not
support translation, all code that referred to this variable was
removed. During this process, all code that referred to the Google
Translate setting was also removed. The next step is to remove the
DUAL_MODE setting from the back-end.