Clean up rendering of Texcavator index page
Cleaned up the view function that renders the index page of the
Texcavator application. The view function was renamed from texcavator
to index (the old index view was removed in a previous step). The view
function mainly sets a number of settings that are used in the
front-end (javascript).
There are still some issues with the index view, such as the use of a
sub site for the production version of Texcavator and a special (fixed)
port number for the development version. These special settings will be
removed, because they make installing Texcavator more complex. It is
not clear why these special settings were added.