Remove export from proxy()
Texcavator cloud export functionality was called via the proxy function
(in services/views.py). Changed url mapping for services/export to
export view function. Moved export functionality to the export view
function and removed export from proxy().

The new view function was decorated with @csrf_exempt, because the
javascript function that calls services/export uses post instead of
get. This has to do with how Dojo handles the result of the
services/export call. This needs to be investigated and solved later.