Add management command to generate word clouds using term vectors
Because terms aggregations do not work for the kb data (the terms data
structure that must be loaded is simply too big to be loaded in
memory), alternatives to generate word clouds are explored. One option
is to 'manually' generate word clouds using the term vectors stored in
ES. A management command was added to generate word clouds based on
term vectors.