Make size of ES chunks a parameter of the command
If the # of documents that is retrieved in an ES request is too large,
the request will time out. In order to be able to experiment with
different sizes and the impact on the performance, the size of ES
chunks was made a parameter of the termvector_wordclouds command. The
size-of-es-chunks parameter is optional.