Remove references to SUB_SITE
A parameter SUB_SITE was used to distinguish between the development
and the production server (at least it appears so). For Texcavator no
SUB_SITE will be used. Therefore most references to SUB_SITE have been
removed. A few references still remain (and should be removed later),
e.g., in request_server_info. It is not clear what this function is
used for.