Add Django management command to gather query terms
Weighted queries consist of tens to hundreds of terms. In order to be
able to test ES performance on weighted queries, a set of terms is
required. A management command was added to gather query terms from
document titles and store them in the database. The next step is to
generate random weighted queries of a certain size.