Add Django management command to generate wordclouds using requests
Generating wordclouds by manually merging termvectors is relatively
slow. Plus, it seems most time is spend by getting the termvectors from
ES. It is not clear why it takes so long, we suspect it might be
parsing the json. To find out, a Django management command was added to
services that does not use the ES python client, but directly makes
http requests to the ES index using the requests library.
The command works for small document sets (<= 100 documents), but gives
an error message for larger sets. Probably, it is trying to send too
much data in the request (=all document ids). This issue must be solved
before performance of parsing text to a dictionary using json can be
tested.