Add query app
An app was added for saving queries. It was decided not to reuse the
old app (lexicon), because the name doesn't make sense, and because it
contains many references to xtas and other things that are no longer
needed. Some functionality that is present in the lexicon app will be
transferred to the query app, such as stop word lists for queries.