Add view to retrieve query details
Added a view that retrieves a query based on a query id. Currently,
only the query and query id are returned. The view needs to be updated
to return all relevant details (including metadata). Also, a check
needs to be added to verify whether the current user is allowed to
retrieve the query.