Add url mapping and view to retrieve a user's queries
An url mapping (query/) and a view (index) was added to retrieve the
saved queries of a given user.
The user for which the queries are retrieved is send as a parameter
of the request. This should be changed to use the Django user
authentication system.
Currently, the queries retrieved are simply serialized to json. This
should be changed to extract relevant parameters, such as which article
types are excluded, etc.