Add DayStatistic model to query app
In order to generate timeline graphs for queries, the DayStatistic
model was included in the lexicon app. Because the lexicon app is being
replaced by the query app, the DayStatistic model was moved to the
query app. The next step is to change the management command
'gatherdaystatistics' to fill the table in the query app instead of in
the lexicon app.