Fix timeline button
The timeline button still called the (old) lexicon app and was
therefore not working anymore. The button now calls the (new) query
app, and the timeline is displayed again. The next step is to refactor
the timeline view function. Also, clicking a bar in the chart results
in an error, because it needs the global parameter
XTAS_MAX_CLOUD_DOCS_WARN, that was previously removed.