Fix error when clicking on bar in timeline chart
Clicking a bar in the timeline graph resulted in an error, because the
JavaScript function called referred to global xtas variables to give a
warning when there were too many documents in a bin. This warning has
been removed. If we need to restrict the number of documents in a
cloud, the warnings can be reinstalled.