Save date range in query
The date range is part of the metadata that must be saved in queries.
The date range is retrieved from the search parameters and saved in the
query. Next, the other metadata (i.e., article types and distributions)
should be saved in the query.
Also, when the query is loaded, the appropriate date range should be
set in the search parameters (and also used for the search).