Fix setting checkboxes for distributions
The checkboxes for distributions in the config panel were not properly
set according to query metadata when a saved query was loaded. Although
the settings were changed, the checkboxes were not, and because the
research function reads the information from the checkboxes and updates
the settings with this information, metadata was not set correctly.
This is now fixed for distributions. The next step is to also fix is
for article types.