Add url mapping and view for deleting queries
Added an url mapping and a view function to delete a query. What is
still missing from the function is a check whether the query belongs to
the current user (and can therefore be deleted by him). Also, the
delete view function is decorated with @csrf_exempt; this should be
handled by sending the csrf token with the http request to delete a
query.