Fix setting metadata checkboxes/getting search parameters
Loading a saved query did not work when the config panel had not been
opened before, because the checkboxes do not yet exist if the panel
hasn't been displayed. Therefore, the load query function
(researchSubmit) now sets both the search parameters in the config
object, as well as the checkboxes (if they exist).