Update saveQuery JavaScript function
The saveQuery JavaScript function was updated so it can also be used
for updating queries in addition to creating new queries. In order to
do so, a new parameter url was added to specify the url that is called
by saveQuery.
saveQuery is also called when creating a new query, this call must be
updated to also pass the appropriate url.