Rename lexicon/StopWord
The lexicon model StopWord was renamed to StopWordOld, because the old
name (StopWord) causes a naming conflict when also added to the query
models (they both have a foreign key to user, and by default, Django
uses the same name for User.stopword_set). Because the lexicon app
will be replaced by the query app, the old StopWord model was renamed.