Add url mapping and view for new login functionality
An url mapping and view were added for the new login functionality. The
old login functionality did not call the Django login function, meaning
that the user was never really logged in. Logging in the user means
that the standard Django methods for user authentication/verification
can be used. This will lead to cleaner and more understandable code.
What still needs to be done: the user interface must call the new login
functionality. Also, when index.html is loaded, the login dialog is
always displayed, regardless of whether a user is already logged in.
This should be fixed.