Fix user authentication for services app
In order to be able to perform searches and other functions of the
services app, a user must be logged in. The user authentication of
relevant views in the services app is now left to the Django
authentication framework. View functions that will be deleted later
(e.g., the proxy view) have not been decorated with the @login_required
decorator, because they are no longer used.