Update tv_cloud view to call celery task for wordcloud generation
The tv_cloud view was updated to call the celery task
generate_tv_word_cloud and return the task id. The next step is to
update the interface to check whether the task is finished and display
the wordcloud if the task is finished. This requires an additional view
function that returns the status of the task and the results if the task
is finished.