Add url mapping and view to check status of celery tasks
And url mapping and view function were added to the services app to
return the status of a celery task given a task id. Currently, only
results for the generate_tv_cloud task can be requested. The view
function should be made generic (i.e., to return the status of any
celery task, regardless of the task that was called). However, because
of a strange import error, and because at the moment there is only one
type of celery task, only results of the generate_tv_cloud task are
returned (details about the error can be found in the view function).
The next step is to change the interface, so it regularly checks whether
generating the cloud data finished.