Fix generate_tv_cloud return type bug
The generate_tv_cloud task returns a dictionary instead of a string, so
calling json.loads() resulted in an error. This bug was not detected
before, because it is required to restart the celery process after
changes are made to tasks, and celery wasn't restarted after the return
type of the task was changed from unicode string to dictionary.