Update interface to pass the minimum word length to the back end
The removal of short words from cloud data is being moved to the back
end. To be able to remove short words, the minimum word length must be
passed to the back end. The interface was updated to pass the minimum
word length of words in the cloud to the back end. The next step is to
pass this value to the generate_tv_cloud task and remove the words
before returning the cloud data.