Update back end to remove short words from cloud data
The minimum word length is passed to the generate_tv_cloud task and the
relevant functions were updated to remove (or, to be more precise, not
to add) words shorter than the minimum word length. This completes the
filtering of short words in word clouds. The next step is to filter stop
words.