Update generate termvector wordcloud to read search parameters
Instead of using a query id to generate a termvector wordcloud, the
tv_wordcloud view reads all relevant parameters from the http request
directly. This means that if the user changed a query after saving it,
the new settings are passed to the tv_wordcloud view (and subsequently
to the celery task). The generate_tv_wordcloud celery task was also
updated to use the new parameters, instead of what is stored for a
query. The query id is still used to retrieve stopwords for the query.
The next step is to update the interface to pass all relevant parameters
when a wordcloud is requested.