Clean up tv_cloud view and generate_tv_cloud task
The code of the tv_cloud view and of the generate_tv_cloud celery task
was cleaned up in order to comply to python style conventions (PEP 8).
Word cloud generation for single document word clouds and word clouds
based on ids (for timeline word clouds) do not work yet. Because this
requires more fixing than just the tv_cloud view and/or the
generate_tv_cloud task, this is not done in this feature branch, but
saved for later.