Fix display single document wordcloud
Because of an error caused by moving saved queries and stopwords from
the lexicon app to the query app, single document wordclouds were no
longer displayed. This error was fixed by removing the code that tried
to get the stopwords from the database. The next step is to move the
removal of stopwords and short words to the back end.