Fix double words in single document wordclouds
In single document wordclouds, words that are used both used in the
title and in the article text, were not added together. This resulted in
words appearing twice in the wordcloud. Also, when exporting the cloud
data, the terms were not ordered by frequency. Both these problems were
fixed by storing terms and counts in a Counter and returning the 100
most common terms.