Update celery task for timeline wordcloud
The celery task for generating wordclouds was updated for timeline
wordclouds. The task now accepts an optional additional parameter
specifying the document ids the cloud should be generated for. Also, the
task sends an additional parameter to the services/es function for
generating the cloud. This parameter specifies whether the cloud is a
timeline cloud or not. This is needed, because a timeline cloud is
displayed in a different panel than a wordcloud for a query.