Fix timeline wordclouds in the interface
The interface was updated, to correctly display timeline wordclouds. For
timeline wordclouds, polling is started once the celery task id has been
returned from the services/cloud request. Subsequently, when the result
is ready, the result is displayed. The next step is to deal with the
error that is returned when the wordcloud generation fails.