Fix display of timeline results
When the user clicks a bar in the timeline, the search results panel
loads the documents in this bar. The function displaying the results
expected an html string, while a json object containing an html string
was send to this function. This resulted in displaying the json object
containing the html string instead of just the html string. The
function was updated to extract the html string from the json object and
display the html string.