Add url mapping and view for query download (export)
Added an url mapping and a view function for downloading (exporting)
query data. The added url mapping and view function create a zipfile
with the query data on the server. When the zipfile is created, the user
receives an email with a link to download the zipfile. The url mapping
and view function for getting the zipfile from the link still needs to
be added.