Add management command to zip query results
The query data download functionality uses a management command
'zipquerydata'. This command was part of the lexicon app and deleted
together with the lexicon app. Therefore, downloading query results did
no longer work. The management command was retrieved and moved to the
query app. Also, the code was cleaned up a bit.