Refactor the termvector word cloud method and celery task
In order to be able to provide information on the progress of the celery
task to create termvector word clouds, the termvector word cloud method
and the celery task were refactored. A method was added to transform a
Counter containing word frequencies into the output the interface uses
to draw a word cloud.