Refactor termvector wordcloud task
For multiple document wordclouds based on a query (the non-timeline
wordclouds), the termvector wordcloud task first gathered all document
ids before merging the termvectors. This is rather expensive for large
document sets (>5000 documents). The task was refactored to gather a
chunk of document ids, merge the termvectors, gather the next chunk,
etc.