Finish mockup of search form
The index template now contains a mockup of the complete search form.
Currently all elements are visible. Some elements will be hidden in some
steps (e.g., the form to save the query should be displayed after the
user clicked search (a query can only be saved after it was performed)).
All elements have been made visible, so that it is easy to make
screenshots to show to the users.