Add wordcloud
Added a directive for displaying wordclouds and show a wordcloud based
on fixed data in the wordcloud tab.

Issues that need to be solved:
- colors (the words are printed in black, however, this makes it harder
to see where one word ends and where another begins. A solution would be
to color the words, but this doesn't look scientific)
- font size (what is the minimum and maximum font size the users find
acceptable?)
- minimum size (width) of the cloud (words in the cloud are clipped if
the cloud is wider than the bootstrap column. The algorithm for placing
the words often places a word (partly) outside the input dimensions.
Currently, this problem is fixed by wrapping the wordcloud in a div with
a fixed width of 1200 pixels)