Add migrations for services and query app
Added migrations for the services and query app. Migrations are used in
Django 1.7.

Djcelery was not added to the apps under migration. It is not clear why
this failed. Possibly it is not required to do so. For now we ignore
djecelery's migrations.

Also there is an error message about the tests. This needs to be fixed
later.