Make the allowed download size dependent on a permission
Users without the permission query.download_many_documents can download only 10k documents instead of 100k documents. This is something that should ideally be managed from the Django admin page using groups.

WARNING. Before deploying, create a group for the users that should
have permission to download 100k documents and add these users to the
group. Then run the migrations, then add the new permission to the
aforementioned group. In this way, the time window in which users are
needlessly limited in their downloading capabilities can be kept as
short as possible.